auto-improve: Phase 2: Remove functions from cai.py and import from cai_lib

Refs damien-robotsix/robotsix-cai#504
