+++ cai.py
@@ -170,573 +170,24 @@ from pathlib import Path
 from publish import ensure_all_labels
 
 
-REPO = "damien-robotsix/robotsix-cai"
-SMOKE_PROMPT = "Say hello in one short sentence."
-
-# Root of claude-code's per-cwd transcript dirs. claude-code writes
-# `~/.claude/projects/<sanitized-cwd>/<session-id>.jsonl` for every
-# session, so this directory contains one subdir per cwd:
-#   * `-app/`            — sessions started by cai.py inside /app
-#   * `-tmp-cai-fix-<N>/` — sessions started by the fix subagent in
-#                          its per-issue clone under /tmp
-# The analyzer parses *all* of them so the fix subagent's tool-rich
-# sessions feed back into the next analyzer cycle.
-#
-# Path is /home/cai/... because the container runs as the non-root
-# `cai` user (uid 1000) — see Dockerfile.
-TRANSCRIPT_DIR = Path("/home/cai/.claude/projects")
-
-# Files baked into the image alongside cai.py.
-PARSE_SCRIPT = Path("/app/parse.py")
-PUBLISH_SCRIPT = Path("/app/publish.py")
-# Persistent memory file for the code-audit agent. Stored in the
-# named-volume log directory so it survives container restarts.
-CODE_AUDIT_MEMORY = Path("/var/log/cai/code-audit-memory.md")
-# Persistent memory file for the propose agent (same pattern).
-PROPOSE_MEMORY = Path("/var/log/cai/propose-memory.md")
-# Persistent memory file for the update-check agent.
-UPDATE_CHECK_MEMORY = Path("/var/log/cai/update-check-memory.md")
-# Persistent memory file for the cost-optimize agent.
-COST_OPTIMIZE_MEMORY = Path("/var/log/cai/cost-optimize-memory.md")
-
-# Persistent per-agent memory directory. Each declarative subagent
-# has `memory: project` in its frontmatter, which Claude Code stores
-# under `.claude/agent-memory/<agent-name>/MEMORY.md` relative to
-# the project root. This directory is bind-mounted from the
-# `cai_agent_memory` named volume so the memory survives container
-# restarts. ALL subagents (both /app agents and the cloned-worktree
-# agents) now read/write this path directly because they're all
-# invoked with `cwd=/app`. The cloned-worktree agents
-# (cai-fix, cai-revise, cai-rebase, cai-review-pr, cai-review-docs, cai-code-audit, cai-propose,
-# cai-propose-review, cai-update-check, cai-plan, cai-select, cai-git) operate
-# on a clone elsewhere via absolute paths —
-# see `_work_directory_block` for the user-message section that
-# tells them where the clone is.
-AGENT_MEMORY_DIR = Path("/app/.claude/agent-memory")
-
-# Issue lifecycle labels.
-LABEL_RAISED = "auto-improve:raised"
-LABEL_REQUESTED = "auto-improve:requested"
-LABEL_IN_PROGRESS = "auto-improve:in-progress"
-LABEL_PR_OPEN = "auto-improve:pr-open"
-LABEL_MERGED = "auto-improve:merged"
-LABEL_SOLVED = "auto-improve:solved"
-LABEL_NO_ACTION = "auto-improve:no-action"
-LABEL_NEEDS_SPIKE = "auto-improve:needs-spike"
-LABEL_NEEDS_EXPLORATION = "auto-improve:needs-exploration"
-LABEL_REFINED = "auto-improve:refined"
-LABEL_REVISING = "auto-improve:revising"
-LABEL_PARENT = "auto-improve:parent"
-LABEL_MERGE_BLOCKED = "merge-blocked"
-LABEL_AUDIT_RAISED = "audit:raised"
-LABEL_AUDIT_NEEDS_HUMAN = "audit:needs-human"
-LABEL_HUMAN_SUBMITTED = "human:submitted"
-LABEL_PLANNED = "auto-improve:planned"
-LABEL_PLAN_APPROVED = "auto-improve:plan-approved"
-
-# PR-level label applied by `cai merge` when the verdict is below the
-# auto-merge threshold. Lets a human filter open PRs that are waiting
-# on their decision (`label:needs-human-review`). Issue #216.
-LABEL_PR_NEEDS_HUMAN = "needs-human-review"
+from cai_lib.config import *  # noqa: E402,F403
 
 
-# ---------------------------------------------------------------------------
-# Run log
-# ---------------------------------------------------------------------------
-
-LOG_PATH = Path("/var/log/cai/cai.log")
-COST_LOG_PATH = Path("/var/log/cai/cai-cost.jsonl")
-REVIEW_PR_PATTERN_LOG = Path("/var/log/cai/review-pr-patterns.jsonl")
-OUTCOME_LOG_PATH = Path("/var/log/cai/cai-outcomes.jsonl")
-ACTIVE_JOB_PATH = Path("/var/log/cai/cai-active.json")
-
-
-def _write_active_job(cmd: str, target_type: str, target_id: int | None) -> None:
-    """Write active-job state for observability. Never raises."""
-    try:
-        ACTIVE_JOB_PATH.parent.mkdir(parents=True, exist_ok=True)
-        ACTIVE_JOB_PATH.write_text(json.dumps({
-            "pid": os.getpid(),
-            "cmd": cmd,
-            "target_type": target_type,
-            "target_id": target_id,
-            "start_ts": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"),
-        }))
-    except OSError:
-        pass
-
-
-def _clear_active_job() -> None:
-    """Clear active-job state file. Never raises."""
-    try:
-        ACTIVE_JOB_PATH.write_text("{}")
-    except OSError:
-        pass
-
-
-def log_run(category: str, **fields) -> None:
-    """Append one key=value line to the persistent run log. Never raises."""
-    try:
-        LOG_PATH.parent.mkdir(parents=True, exist_ok=True)
-        ts = datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ")
-        parts = [f"{ts} [{category}]"]
-        for k, v in fields.items():
-            parts.append(f"{k}={v}")
-        line = " ".join(parts) + "\n"
-        with LOG_PATH.open("a") as f:
-            f.write(line)
-            f.flush()
-    except Exception:
-        pass
-
-
-def log_cost(row: dict) -> None:
-    """Append one JSON object to the per-invocation cost log. Never raises.
-
-    Each row records the cost and token usage of a single `claude -p`
-    invocation, plus the cai-side context (category, agent) so the
-    audit agent and the `cost-report` subcommand can attribute spend
-    to specific cai commands and subagents.
-    """
-    try:
-        COST_LOG_PATH.parent.mkdir(parents=True, exist_ok=True)
-        with COST_LOG_PATH.open("a") as f:
-            f.write(json.dumps(row, separators=(",", ":")) + "\n")
-            f.flush()
-    except Exception:
-        pass
-
-
-def _get_issue_category(issue: dict) -> str:
-    """Return the category label value for *issue*, or ``'(unknown)'`` if absent."""
-    for ln in (lbl["name"] for lbl in issue.get("labels", [])):
-        if ln.startswith("category:"):
-            return ln.split(":", 1)[1]
-    return "(unknown)"
-
-
-def _log_outcome(issue_number: int, category: str, outcome: str, fix_attempt_count: int) -> None:
-    """Append one JSON record to the outcome log. Never raises."""
-    try:
-        OUTCOME_LOG_PATH.parent.mkdir(parents=True, exist_ok=True)
-        row = {
-            "ts": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"),
-            "issue_number": issue_number,
-            "category": category,
-            "outcome": outcome,
-            "fix_attempt_count": fix_attempt_count,
-        }
-        with OUTCOME_LOG_PATH.open("a") as f:
-            f.write(json.dumps(row, separators=(",", ":")) + "\n")
-            f.flush()
-    except Exception:
-        pass
-
-
-def _load_outcome_counts(days: int = 90) -> dict:
-    """Read OUTCOME_LOG_PATH and return per-category {total, solved} counts.
-
-    Filters to trailing `days` days. Malformed lines are skipped silently.
-    Returns an empty dict if the file is missing or unreadable.
-    """
-    if not OUTCOME_LOG_PATH.exists():
-        return {}
-    cutoff_ts = datetime.now(timezone.utc).timestamp() - days * 86400
-    counts: dict = {}  # category -> {"total": N, "solved": N}
-    try:
-        with OUTCOME_LOG_PATH.open("r") as f:
-            for line in f:
-                line = line.strip()
-                if not line:
-                    continue
-                try:
-                    row = json.loads(line)
-                except (json.JSONDecodeError, ValueError):
-                    continue
-                ts = row.get("ts", "")
-                try:
-                    row_ts = datetime.strptime(
-                        ts, "%Y-%m-%dT%H:%M:%SZ"
-                    ).replace(tzinfo=timezone.utc).timestamp()
-                except ValueError:
-                    continue
-                if row_ts < cutoff_ts:
-                    continue
-                cat = row.get("category") or "(unknown)"
-                outcome = row.get("outcome", "")
-                bucket = counts.setdefault(cat, {"total": 0, "solved": 0})
-                bucket["total"] += 1
-                if outcome == "solved":
-                    bucket["solved"] += 1
-    except OSError:
-        return {}
-    return counts
-
-
-def _load_outcome_stats(days: int = 90) -> dict:
-    """Load per-category success rates from the trailing `days` days of outcome data.
-
-    Returns a dict mapping category name to success rate (0.0–1.0).
-    Categories with fewer than 3 observations get a neutral prior of 0.60.
-    """
-    counts = _load_outcome_counts(days)
-    rates: dict = {}
-    for cat, c in counts.items():
-        if c["total"] < 3:
-            rates[cat] = 0.60
-        else:
-            rates[cat] = c["solved"] / c["total"]
-    return rates
-
-
-def _load_cost_log(days: int = 7) -> list[dict]:
-    """Read COST_LOG_PATH and return rows from the last `days` days.
-
-    Each row is a dict as written by `log_cost`. Malformed lines are
-    skipped silently. Returns an empty list if the file is missing or
-    unreadable. Used by both `_build_cost_summary` (audit prompt) and
-    `cmd_cost_report` (host-facing report).
-    """
-    if not COST_LOG_PATH.exists():
-        return []
-    cutoff_ts = datetime.now(timezone.utc).timestamp() - days * 86400
-    rows: list[dict] = []
-    try:
-        with COST_LOG_PATH.open("r") as f:
-            for line in f:
-                line = line.strip()
-                if not line:
-                    continue
-                try:
-                    row = json.loads(line)
-                except (json.JSONDecodeError, ValueError):
-                    continue
-                ts = row.get("ts") or ""
-                try:
-                    row_ts = datetime.strptime(
-                        ts, "%Y-%m-%dT%H:%M:%SZ",
-                    ).replace(tzinfo=timezone.utc).timestamp()
-                except ValueError:
-                    continue
-                if row_ts >= cutoff_ts:
-                    rows.append(row)
-    except Exception:
-        return []
-    return rows
-
-
-def _row_ts(row: dict) -> float:
-    """Parse a cost-log row's 'ts' field to a Unix timestamp.
-
-    Returns 0.0 on any parse failure so callers can safely compare
-    against numeric boundaries without extra error handling.
-    """
-    ts = row.get("ts") or ""
-    try:
-        return datetime.strptime(
-            ts, "%Y-%m-%dT%H:%M:%SZ",
-        ).replace(tzinfo=timezone.utc).timestamp()
-    except ValueError:
-        return 0.0
-
-
-def _build_cost_summary(days: int = 7, top_n: int = 10) -> str:
-    """Build a markdown cost summary for the cai-audit user message.
-
-    Returns an empty string if no cost rows exist for the window.
-    Otherwise emits a section with per-category aggregates and the
-    top-N most expensive individual invocations, so the audit agent
-    can spot cost outliers (a single invocation that dwarfs the
-    median, or a category that dominates total spend).
-    """
-    rows = _load_cost_log(days=days)
-    if not rows:
-        return ""
-
-    # Per-category aggregates: total cost, call count, mean cost.
-    cats: dict[str, dict] = {}
-    grand_total = 0.0
-    for r in rows:
-        cat = r.get("category") or "(unknown)"
-        cost = r.get("cost_usd") or 0.0
-        try:
-            cost = float(cost)
-        except (TypeError, ValueError):
-            cost = 0.0
-        bucket = cats.setdefault(cat, {"calls": 0, "cost": 0.0})
-        bucket["calls"] += 1
-        bucket["cost"] += cost
-        grand_total += cost
-
-    cat_lines = []
-    for cat, b in sorted(cats.items(), key=lambda kv: -kv[1]["cost"]):
-        share = (b["cost"] / grand_total * 100.0) if grand_total else 0.0
-        mean = b["cost"] / b["calls"] if b["calls"] else 0.0
-        cat_lines.append(
-            f"| {cat} | {b['calls']} | ${b['cost']:.4f} "
-            f"({share:.1f}%) | ${mean:.4f} |"
-        )
-
-    # Top-N most expensive individual invocations.
-    top = sorted(
-        rows,
-        key=lambda r: float(r.get("cost_usd") or 0.0),
-        reverse=True,
-    )[:top_n]
-    top_lines = []
-    for r in top:
-        cost = float(r.get("cost_usd") or 0.0)
-        top_lines.append(
-            f"| {r.get('ts', '')} | {r.get('category', '')} | "
-            f"{r.get('agent', '')} | ${cost:.4f} | "
-            f"{r.get('num_turns', '')} | "
-            f"{(r.get('input_tokens') or 0) + (r.get('output_tokens') or 0)} |"
-        )
-
-    return (
-        f"## Cost summary (last {days}d, total ${grand_total:.4f} "
-        f"across {len(rows)} invocations)\n\n"
-        "### Per-category totals\n\n"
-        "| category | calls | total cost (share) | mean cost |\n"
-        "|---|---|---|---|\n"
-        + "\n".join(cat_lines)
-        + "\n\n"
-        f"### Top {len(top_lines)} most expensive individual invocations\n\n"
-        "| ts | category | agent | cost | turns | tokens |\n"
-        "|---|---|---|---|---|---|\n"
-        + "\n".join(top_lines)
-        + "\n"
-    )
+from cai_lib.logging_utils import (  # noqa: E402
+    _write_active_job, _clear_active_job, log_run, log_cost,
+    _get_issue_category, _log_outcome, _load_outcome_counts,
+    _load_outcome_stats, _load_cost_log, _row_ts, _build_cost_summary,
+)
 
 
 # ---------------------------------------------------------------------------
 # Common helpers
 # ---------------------------------------------------------------------------
 
-def _run(cmd: list[str], **kwargs) -> subprocess.CompletedProcess:
-    """Thin wrapper around subprocess.run with text mode and check=False."""
-    return subprocess.run(cmd, text=True, check=False, **kwargs)
+from cai_lib.subprocess_utils import _run, _run_claude_p  # noqa: E402
 
 
-def _run_claude_p(
-    cmd: list[str],
-    *,
-    category: str,
-    agent: str = "",
-    **kwargs,
-) -> subprocess.CompletedProcess:
-    """Run a `claude -p` command and record its cost.
-
-    `cmd` is the full argv. The wrapper injects `--output-format json
-    --verbose` so claude-code returns the cost/usage bookkeeping for
-    the run. With `--verbose`, claude-code emits a JSON **array** of
-    stream events (`system` → `assistant` → `user` → `result`); the
-    `result` element holds `total_cost_usd`, `usage`, `duration_ms`,
-    `result` text, etc. We extract that element, log a cost row, and
-    rewrite `CompletedProcess.stdout` to just the `result` text — so
-    existing callers that pipe `proc.stdout` to `publish.py` or print
-    it keep working unchanged.
-
-    `category` labels the row by top-level cai command (e.g.
-    "analyze", "fix", "audit"). `agent` records the subagent name
-    (e.g. "cai-fix") if applicable.
-
-    On JSON parse failure or a missing `result` event, no cost row is
-    written, the original stdout is left in place, and a one-line
-    warning is printed to stderr so this silent-drop failure mode is
-    noisy. Never raises.
-    """
-    # Inject --output-format json --verbose right after `claude -p`
-    # (positions 0 and 1). --verbose is required for claude-code to
-    # populate the `usage` field; with it, the output becomes a JSON
-    # array of stream events instead of a single envelope dict.
-    if len(cmd) < 2 or cmd[0] != "claude" or cmd[1] != "-p":
-        raise ValueError("_run_claude_p requires cmd[:2] == ['claude', '-p']")
-    plugin_dir = Path(".claude/plugins/cai-skills")
-    plugin_flags: list[str] = (
-        ["--plugin-dir", str(plugin_dir)] if plugin_dir.is_dir() else []
-    )
-    full_cmd = (
-        cmd[:2]
-        + ["--output-format", "json", "--verbose"]
-        + plugin_flags
-        + cmd[2:]
-    )
-
-    # Force capture so we can parse the JSON envelope. Callers that
-    # previously did not capture (only cmd_init) get back the result
-    # text in `.stdout` — they can print it themselves if needed.
-    kwargs.setdefault("capture_output", True)
-    proc = _run(full_cmd, **kwargs)
-
-    # Parse the JSON envelope and write the cost row. Belt and braces
-    # — never let log writes break the actual command flow.
-    try:
-        parsed = json.loads(proc.stdout) if proc.stdout else None
-    except (json.JSONDecodeError, ValueError):
-        parsed = None
-
-    # Two shapes are tolerated:
-    #   1. dict   — legacy `--output-format json` (no --verbose) returns
-    #      a single envelope object. Kept for forward/backward compat.
-    #   2. list   — current `--output-format json --verbose` returns a
-    #      JSON array of stream events; the cost data lives on the
-    #      element with `"type": "result"`.
-    envelope: dict | None = None
-    subagent_results: list[dict] = []
-    if isinstance(parsed, dict):
-        envelope = parsed
-    elif isinstance(parsed, list):
-        result_events = [
-            e for e in parsed
-            if isinstance(e, dict) and e.get("type") == "result"
-        ]
-        # The last result event is the parent (top-level) result;
-        # earlier ones are subagent results.
-        envelope = result_events[-1] if result_events else None
-        subagent_results = result_events[:-1] if len(result_events) > 1 else []
-
-    if envelope is None:
-        # Don't fail the caller, but make the silent-drop loud so a
-        # future shape change in claude-code surfaces immediately
-        # instead of leaving cai-cost.jsonl mysteriously empty.
-        preview = (proc.stdout or "")[:120].replace("\n", " ")
-        print(
-            f"[cai cost] could not extract cost envelope from claude -p "
-            f"({category}/{agent}); stdout starts with: {preview!r}",
-            file=sys.stderr,
-            flush=True,
-        )
-
-    if isinstance(envelope, dict):
-        usage = envelope.get("usage") or {}
-        # claude-code's `usage` may be either a flat dict (input_tokens,
-        # output_tokens, cache_*_input_tokens) or a nested per-model
-        # dict. Record both shapes when available.
-        flat_keys = (
-            "input_tokens",
-            "output_tokens",
-            "cache_creation_input_tokens",
-            "cache_read_input_tokens",
-        )
-        flat = {k: usage[k] for k in flat_keys if isinstance(usage.get(k), (int, float))}
-        models = {
-            k: v for k, v in usage.items()
-            if isinstance(v, dict) and any(fk in v for fk in flat_keys)
-        }
-
-        # -- Subagent token aggregation --
-        subagent_rows: list[dict] = []
-        combined = dict(flat)  # start with parent tokens
-        for sr in subagent_results:
-            sr_usage = sr.get("usage") or {}
-            sr_flat = {k: sr_usage[k] for k in flat_keys if isinstance(sr_usage.get(k), (int, float))}
-            if sr_flat:
-                for k in flat_keys:
-                    if k in sr_flat:
-                        combined[k] = combined.get(k, 0) + sr_flat[k]
-                sr_entry: dict = dict(sr_flat)
-                sr_entry["cost_usd"] = sr.get("total_cost_usd")
-                subagent_rows.append(sr_entry)
-
-        row = {
-            "ts": datetime.now(timezone.utc).strftime("%Y-%m-%dT%H:%M:%SZ"),
-            "category": category,
-            "agent": agent,
-            "cost_usd": envelope.get("total_cost_usd"),
-            "duration_ms": envelope.get("duration_ms"),
-            "duration_api_ms": envelope.get("duration_api_ms"),
-            "num_turns": envelope.get("num_turns"),
-            "session_id": envelope.get("session_id"),
-            "exit": proc.returncode,
-            "is_error": bool(envelope.get("is_error", proc.returncode != 0)),
-        }
-        row.update(combined)
-        if models:
-            row["models"] = models
-        if subagent_rows:
-            sub_cost_sum = sum(float(s.get("cost_usd") or 0.0) for s in subagent_rows)
-            total = float(envelope.get("total_cost_usd") or 0.0)
-            row["parent_cost_usd"] = round(total - sub_cost_sum, 6)
-            row["subagents"] = subagent_rows
-        log_cost(row)
-
-        # Rewrite stdout to the result text so existing callers stay
-        # backwards compatible. If `result` is missing, fall back to
-        # the raw envelope so callers still see *something*.
-        if "result" in envelope and isinstance(envelope["result"], str):
-            proc.stdout = envelope["result"]
-
-    return proc
-
-
-def _gh_json(args: list[str]):
-    """Run a gh command that prints JSON; return the parsed result.
-
-    Raises on non-zero exit (gh failures should be loud).
-    """
-    result = subprocess.run(
-        ["gh"] + args,
-        text=True,
-        check=True,
-        capture_output=True,
-    )
-    return json.loads(result.stdout) if result.stdout.strip() else None
-
-
-def check_gh_auth() -> int:
-    """Fail fast if `gh` is not authenticated."""
-    result = _run(["gh", "auth", "status"], capture_output=True)
-    if result.returncode != 0:
-        print("[cai] ERROR: gh is not authenticated in this container.", file=sys.stderr)
-        print("       Credentials are expected in the cai_home volume.", file=sys.stderr)
-        print("       Run the installer's login step, or do it manually:", file=sys.stderr)
-        print("         docker compose run --rm cai gh auth login", file=sys.stderr)
-        print(file=sys.stderr)
-        print(result.stderr.strip() or result.stdout.strip(), file=sys.stderr)
-        return 1
-    return 0
-
-
-def check_claude_auth() -> int:
-    """Fail fast if `claude` is not authenticated.
-
-    Two valid auth modes for the headless container:
-      1. OAuth: credentials sit in the `cai_home` named volume
-         (under `/home/cai/.claude/.credentials.json` plus the
-         `/home/cai/.claude.json` runtime config sibling file).
-         Verified by `claude auth status`.
-      2. API key: `ANTHROPIC_API_KEY` is set in the env. claude-code
-         uses it directly without needing the OAuth credentials file.
-
-    If neither mode is configured, claude-code will 401 on the first
-    API call with a confusing error. Catch the misconfiguration up
-    front so the user gets a clear next-step instruction.
-    """
-    # API-key mode is checked first because it's a single env-var test
-    # and doesn't require shelling out.
-    if os.environ.get("ANTHROPIC_API_KEY"):
-        return 0
-
-    result = _run(["claude", "auth", "status", "--text"], capture_output=True)
-    if result.returncode != 0:
-        print("[cai] ERROR: claude is not authenticated in this container.", file=sys.stderr)
-        print("       Credentials are expected in the cai_home volume,", file=sys.stderr)
-        print("       OR set ANTHROPIC_API_KEY in your .env file.", file=sys.stderr)
-        print("       Authenticate by opening the claude REPL — it auto-prompts", file=sys.stderr)
-        print("       for OAuth login on first start:", file=sys.stderr)
-        print("         docker compose run --rm -it cai claude", file=sys.stderr)
-        print("       Then exit the REPL gracefully (/exit or Ctrl-D).", file=sys.stderr)
-        print(file=sys.stderr)
-        print(result.stderr.strip() or result.stdout.strip(), file=sys.stderr)
-        return 1
-    return 0
-
-
-def _transcript_dir_is_empty() -> bool:
-    if not TRANSCRIPT_DIR.exists():
-        return True
-    return not any(TRANSCRIPT_DIR.rglob("*.jsonl"))
+from cai_lib.github import _gh_json, check_gh_auth, check_claude_auth, _transcript_dir_is_empty  # noqa: E402
 
 
 # ---------------------------------------------------------------------------
@@ -1343,75 +794,7 @@ def _select_plan_target(issue_number: int | None = None):
     return min(candidates, key=lambda c: c.get("createdAt", ""))
 
 
-def _set_labels(issue_number: int, *, add: list[str] = (), remove: list[str] = (), log_prefix: str = "cai fix") -> bool:
-    """Add and/or remove labels on an issue. Returns True on success."""
-    # Auto-add the base label for any state-prefixed label being added.
-    # This is defensive: create_issue already applies base labels, but
-    # auto-adding here self-heals issues that lost theirs.
-    _BASE_NAMESPACES = {"auto-improve", "audit", "check-workflows"}
-    auto_added_bases: set[str] = set()
-    for label in add:
-        if ":" in label:
-            base = label.split(":", 1)[0]
-            if base in _BASE_NAMESPACES and base not in add:
-                auto_added_bases.add(base)
-    effective_add = list(add) + sorted(auto_added_bases)
-
-    args = ["issue", "edit", str(issue_number), "--repo", REPO]
-    for label in effective_add:
-        args.extend(["--add-label", label])
-    for label in remove:
-        args.extend(["--remove-label", label])
-    result = _run(["gh"] + args, capture_output=True)
-    if result.returncode != 0:
-        print(
-            f"[{log_prefix}] failed to update labels on #{issue_number}:\n{result.stderr}",
-            file=sys.stderr,
-        )
-        return False
-    return True
-
-
-def _issue_has_label(issue_number: int, label: str) -> bool:
-    """Re-fetch an issue's labels and check for *label*. Avoids stale-snapshot races."""
-    try:
-        issue = _gh_json([
-            "issue", "view", str(issue_number),
-            "--repo", REPO,
-            "--json", "labels",
-        ])
-    except subprocess.CalledProcessError:
-        return False
-    return label in [l["name"] for l in (issue or {}).get("labels", [])]
-
-
-def _build_issue_block(issue: dict) -> str:
-    """Build the issue block shared by plan, select, and fix agents."""
-    block = (
-        f"## Issue\n\n"
-        f"### #{issue['number']} — {issue['title']}\n\n"
-        f"{issue.get('body') or '(no body)'}\n"
-    )
-    comments = issue.get("comments") or []
-    if comments:
-        block += "\n### Comments\n\n"
-        for c in comments:
-            author = c.get("author", {}).get("login", "unknown")
-            body = c.get("body", "")
-            block += f"**{author}:**\n{body}\n\n"
-    return block
-
-
-def _build_fix_user_message(issue: dict, attempt_history_block: str = "") -> str:
-    """Build the dynamic per-run user message for the cai-fix agent.
-
-    The system prompt, tool allowlist, and hard rules live in
-    `.claude/agents/cai-fix.md`; durable per-agent learnings live
-    in its `memory: project` pool. The wrapper passes the issue
-    body, reviewer comments, and (when available) a summary of
-    prior closed PRs for this issue.
-    """
-    return _build_issue_block(issue) + attempt_history_block
+from cai_lib.github import _set_labels, _issue_has_label, _build_issue_block, _build_fix_user_message  # noqa: E402
 
 
 def _fetch_previous_fix_attempts(issue_number: int) -> list[dict]:
@@ -1811,47 +1194,7 @@ def _create_suggested_issues(
 # ---------------------------------------------------------------------------
 
 
-def _parse_decomposition(agent_output: str) -> list[dict]:
-    """Extract ordered steps from a ``## Multi-Step Decomposition`` block.
-
-    Expected format in *agent_output*::
-
-        ## Multi-Step Decomposition
-
-        ### Step 1: <title>
-        <body>
-
-        ### Step 2: <title>
-        <body>
-
-    Returns a list of ``{"step": int, "title": str, "body": str}`` dicts,
-    sorted by step number.  Returns an empty list when the marker is
-    missing or the output is malformed.
-    """
-    marker = "## Multi-Step Decomposition"
-    marker_pos = agent_output.find(marker)
-    if marker_pos == -1:
-        return []
-
-    text = agent_output[marker_pos + len(marker):]
-    parts = re.split(r"^### Step (\d+):\s*", text, flags=re.MULTILINE)
-    # parts[0] is preamble (before first step), then alternating
-    # (step_number, body) pairs.
-    steps: list[dict] = []
-    i = 1
-    while i + 1 < len(parts):
-        step_num = int(parts[i])
-        raw = parts[i + 1].strip()
-        # The title is the first non-empty line; the rest is the body.
-        lines = raw.split("\n", 1)
-        title = lines[0].strip()
-        body = lines[1].strip() if len(lines) > 1 else ""
-        if title:
-            steps.append({"step": step_num, "title": title, "body": body})
-        i += 2
-
-    steps.sort(key=lambda s: s["step"])
-    return steps
+from cai_lib.cmd_fix import _parse_decomposition  # noqa: E402
 
 
 def _find_sub_issue(parent_number: int, step: int) -> int | None:
@@ -4098,10 +3441,7 @@ def cmd_verify(args) -> int:
 # audit
 # ---------------------------------------------------------------------------
 
-_STALE_IN_PROGRESS_HOURS = 6
-_STALE_REVISING_HOURS = 1
-_STALE_NO_ACTION_DAYS = 7
-_STALE_MERGED_DAYS = 14
+from cai_lib.config import _STALE_IN_PROGRESS_HOURS, _STALE_REVISING_HOURS, _STALE_NO_ACTION_DAYS, _STALE_MERGED_DAYS  # noqa: E402
 
 
 
@@ -4183,126 +3523,7 @@ def _cleanup_orphaned_branches() -> list[str]:
     return deleted
 
 
-def _rollback_stale_in_progress(*, immediate: bool = False) -> list[dict]:
-    """Deterministic rollback: :in-progress or :revising issues with no recent activity.
-
-    When ``immediate=True`` every locked issue is rolled back regardless of age
-    (used by ``cmd_cycle`` on container restart where all in-flight locks are
-    guaranteed to be orphaned).
-
-    Returns the list of issues that were rolled back.
-    """
-    all_issues = []
-    for lock_label in (LABEL_IN_PROGRESS, LABEL_REVISING):
-        try:
-            issues = _gh_json([
-                "issue", "list",
-                "--repo", REPO,
-                "--label", lock_label,
-                "--state", "open",
-                "--json", "number,title,updatedAt,createdAt,labels",
-                "--limit", "100",
-            ]) or []
-        except subprocess.CalledProcessError as e:
-            print(
-                f"[cai audit] gh issue list ({lock_label}) failed:\n{e.stderr}",
-                file=sys.stderr,
-            )
-            continue
-        for issue in issues:
-            issue["_lock_label"] = lock_label
-            all_issues.append(issue)
-
-    if not all_issues:
-        return []
-
-    issues = all_issues
-
-    # Read the log tail to find the most recent [fix] line per issue.
-    fix_timestamps: dict[int, float] = {}
-    if LOG_PATH.exists():
-        try:
-            lines = LOG_PATH.read_text().splitlines()[-200:]
-        except Exception:
-            lines = []
-        for line in lines:
-            if "[fix]" not in line and "[revise]" not in line and "[spike]" not in line:
-                continue
-            # Extract issue number from "issue=<N>"
-            m = re.search(r"issue=(\d+)", line)
-            if not m:
-                continue
-            issue_num = int(m.group(1))
-            # Extract timestamp from start of line (ISO format)
-            ts_match = re.match(r"(\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}Z)", line)
-            if ts_match:
-                try:
-                    ts = datetime.strptime(ts_match.group(1), "%Y-%m-%dT%H:%M:%SZ").replace(
-                        tzinfo=timezone.utc
-                    ).timestamp()
-                    fix_timestamps[issue_num] = max(fix_timestamps.get(issue_num, 0), ts)
-                except ValueError:
-                    pass
-
-    now = datetime.now(timezone.utc).timestamp()
-    rolled_back = []
-
-    for issue in issues:
-        issue_num = issue["number"]
-        lock_label = issue.get("_lock_label", LABEL_IN_PROGRESS)
-        ttl_hours = _STALE_REVISING_HOURS if lock_label == LABEL_REVISING else _STALE_IN_PROGRESS_HOURS
-        threshold = 0 if immediate else ttl_hours * 3600
-        last_fix = fix_timestamps.get(issue_num)
-        if last_fix is not None:
-            age = now - last_fix
-        else:
-            # No fix log line — use the issue's updatedAt as a fallback.
-            try:
-                updated = datetime.strptime(
-                    issue["updatedAt"], "%Y-%m-%dT%H:%M:%SZ"
-                ).replace(tzinfo=timezone.utc).timestamp()
-            except (ValueError, KeyError):
-                updated = 0
-            age = now - updated
-
-        if age > threshold:
-            if lock_label == LABEL_REVISING:
-                # Revising lock: just remove the lock, leave :pr-open.
-                ok = _set_labels(issue_num, remove=[LABEL_REVISING], log_prefix="cai audit")
-            else:
-                # In-progress lock: roll back to the appropriate label.
-                # Check originating label: spike-provenance issues go back to
-                # :needs-spike; audit-raised go back to :audit-raised; all
-                # others go back to :refined.
-                issue_labels = {lbl["name"] for lbl in issue.get("labels", [])}
-                if LABEL_AUDIT_RAISED in issue_labels:
-                    raised_label = LABEL_AUDIT_RAISED
-                elif LABEL_NEEDS_SPIKE in issue_labels:
-                    raised_label = LABEL_NEEDS_SPIKE
-                else:
-                    raised_label = LABEL_REFINED
-                ok = _set_labels(
-                    issue_num,
-                    add=[raised_label],
-                    remove=[LABEL_IN_PROGRESS],
-                    log_prefix="cai audit",
-                )
-            if ok:
-                rolled_back.append(issue)
-                log_run(
-                    "audit",
-                    action="stale_lock_rollback",
-                    issue=issue_num,
-                    lock_label=lock_label,
-                    stale_hours=f"{age / 3600:.1f}",
-                )
-                print(
-                    f"[cai audit] rolled back #{issue_num} "
-                    f"(removed {lock_label}, stale {age / 3600:.1f}h)",
-                    flush=True,
-                )
-
-    return rolled_back
+from cai_lib.cmd_lifecycle import _rollback_stale_in_progress  # noqa: E402
 
 
 def _unstuck_stale_no_action() -> list[dict]:
